requirements.txt
Django==1.7.2
dj-database-url==0.3.0
dj-static==0.0.6
gunicorn==19.1.1
psycopg2==2.5.4
static3==0.5.1
python-social-auth==0.2.1
